Created EigenSparseHelpers
EigenHelpers do not contain Sparse code.
It is possible then to use EigenHelpers on Eigen < 3.3 and
EigenSparseHelpers on Eigen >= 3.3
