+++ src/core/tests/CMakeLists.txt
@@ -25,3 +25,4 @@ add_unit_test(VectorDynSize)
 add_unit_test(MatrixDynSize)
 add_unit_test(SparseMatrix) 
 add_unit_test(EigenHelpers)
+add_unit_test(EigenSparseHelpers)
